Bug with different gpus
Even though the code is deterministic, it is seemingly random whether things will render or not. My hunch is that the GPU on the current computer is damaged (I was told so by the previous owner). This explanation does not make the most sense, but it is the easiest one to test and could very well be the problem.