Tidied code
There is no reason this code should not be working. The buffers are properly filled, bound, and drawn with the correct arguments. Yet sometimes, WGPU just decides it only wants to render 1 of them, or 2 or 3 sometimes, just depending on how it feels. The bug comes and goes with no rhyme or reason. My belief is that it may be a problem with WGPU itself, or something other than my code. I will continue to investigate.